Rename test package to allow usage of package functions
